Inline (but don't yet simplify) S.Context,Session
